Narrow the source location offsets to 32-bit fields
One shell source is far below four gigabytes, so a 64-bit byte offset
wastes half its width. Every token and every syntax node carries one
SourceLocation, so the pair is on the hottest memory path there is.

Changes:

1. SourceLocation::position and SourceLocation::length are u32. A new
   constructor accepts usize, so the many call sites that compute an
   offset as usize compile unchanged.
2. Six direct field assignments in Cli.cpp, Eval.cpp, Expressions.cpp,
   ExpressionsArith.cpp, ExpressionsSimpleCommand.cpp, and
   ParserCompound.cpp narrow explicitly.
3. Ten parser log messages read the offset with %u.

SourceLocation is 32 bytes rather than 40, Token is 40 rather than 48,
WordToken is 224 rather than 232, and SimpleCommand is 256 rather than
272. Peak resident memory on a 15 MB configure script falls from
652 MB to 634 MB.

+++ src/Cli.cpp
@@ -306,8 +306,8 @@ static fn attached_flag_value_location(
       argument_location(argv, argument_index, base_position, arg_locations);
   if (location.length != std::strlen(argv[argument_index])) return location;
 
-  location.position += static_cast<usize>(value - argv[argument_index]);
-  location.length = std::strlen(value);
+  location.position += static_cast<u32>(value - argv[argument_index]);
+  location.length = static_cast<u32>(std::strlen(value));
   return location;
 }
